project/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.0.1
pytesseract==0.3.10
pdf2image==1.16.3
streamlit==1.28.1
requests==2.31.0
python-multipart==0.0.6
google-generativeai==0.3.2